Explicit file move part 1
